2.0: minor updates

+++ src/utils/xpnwriter.c
@@ -80,7 +80,7 @@ int main(int argc, char *argv[])
 #ifdef DEBUG
 	printf("xpndest=%d, isxpn=%d\n", xpndest, isxpn);
 	printf("dest = '%s'\n", dest);
-	printf("buffer_size = %lu\n", buffer_size);
+	printf("buffer_size = %lu\n", (unsigned long)buffer_size);
 #endif
 	if (isxpn) {
 		ret = xpn_init();
